# docs/swagger.yaml
definitions:
...
  model.Account:
    properties:
      ERC:
        description: ERC types, ERC20, ERC721, ERC1155
        type: integer
      address:
        description: address
        type: string
      balance:
        description: The total amount of coins in the chain
        type: string
      code:
        description: bytecode
        type: string
      createdTx:
        description: Create transaction
        type: string
      creator:
        description: address
        type: string
      name:
        description: name
        type: string
      symbol:
        description: symbol
        type: string
      transactionCount:
        description: Transaction random number, transaction volume
        type: integer
    type: object
  model.Block:
    properties:
      difficulty:
        description: Transaction random number, transaction volume
        type: integer
      extraData:
        description: Extra data
        type: string
      gasLimit:
        description: Transaction random number, transaction volume
        type: integer
      gasUsed:
        description: Transaction random number, transaction volume
        type: integer
      hash:
        description: Create transaction
        type: string
      miner:
        description: address
        type: string
      mixHash:
        description: Create transaction
        type: string
      nonce:
        description: difficulty random number
        type: string
      number:
        description: Transaction random number, transaction volume
        type: integer
      parentHash:
        description: Create transaction
        type: string
      receiptsRoot:
        description: Create transaction
        type: string
      sha3Uncles:
        description: Create transaction
        type: string
      size:
        description: Transaction random number, transaction volume
        type: integer
      stateRoot:
        description: Create transaction
        type: string
      timestamp:
        description: Transaction random number, transaction volume
        type: integer
      totalDifficulty:
        description: The total amount of coins in the chain
        type: string
      totalTransaction:
        description: Transaction random number, transaction volume
        type: integer
      transactionsRoot:
        description: Create transaction
        type: string
      uncles:
        description: Uncle block hash
        items:
          type: string
        type: array
      unclesCount:
        description: Transaction random number, transaction volume
        type: integer
    type: object
  model.Collection:
    properties:
      block_number:
        description: Create block height, equal to the first NFT in the collection
        type: integer
      category:
        description: category
        type: string
      creator:
        description: Creator
        type: string
      desc:
        description: description
        type: string
      exchanger:
        description: belongs to the exchange
        type: string
      id:
        description: ID
        type: string
      img_url:
        description: image link
        type: string
      meta_url:
        description: collection meta information URL
        type: string
      name:
        description: name
        type: string
    type: object
...
  service.AccountRes:
    properties:
      ERC:
        description: ERC types, ERC20, ERC721, ERC1155
        type: integer
      address:
        description: address
        type: string
      balance:
        description: The total amount of coins in the chain
        type: string
      code:
        description: bytecode
        type: string
      createdTx:
        description: Create transaction
        type: string
      creator:
        description: address
        type: string
      name:
        description: name
        type: string
      pledgeAmount:
        description: pledge amount
        type: string
      rewardCoinCount:
        description: Number of times to get coin rewards, 0.1ERB once
        type: integer
      rewardSNFTCount:
        description: Number of times to get SNFT rewards
        type: integer
      symbol:
        description: symbol
        type: string
      totalNFT:
        description: Number of NFTs held
        type: integer
      totalSNFT:
        description: Number of SNFTs held
        type: integer
      transactionCount:
        description: Transaction random number, transaction volume
        type: integer
    type: object
  service.AccountsRes:
    properties:
      accounts:
        description: Account list
        items:
          $ref: '#/definitions/model.Account'
        type: array
      balance:
        description: The total amount of coins in the chain
        type: string
      total:
        description: Total number of accounts
        type: integer
    type: object
...
  service.BlocksRes:
    properties:
      blocks:
        description: block list
        items:
          $ref: '#/definitions/model.Block'
        type: array
      total:
        description: The total number of blocks
        type: integer
    type: object
  service.Cache:
    properties:
      genesisBalance:
        description: The total amount of coins in the chain
        type: string
      rewardCoinCount:
        description: Total number of times to get coin rewards, 0.1ERB once
        type: integer
      rewardSNFTCount:
        description: Total number of times to get SNFT rewards
        type: integer
      total24HExchangerTx:
        description: Total number of exchanger  transactions within 24 hours
        type: integer
      total24HNFT:
        description: Total number of NFT within 24 hours
        type: integer
      total24HTx:
        description: Total number of transactions within 24 hours
        type: integer
      totalAccount:
        description: Total account number
        type: integer
      totalBalance:
        description: The total amount of coins in the chain
        type: string
      totalBlock:
        description: Total number of blocks
        type: integer
      totalExchanger:
        description: Total number of exchanges
        type: integer
      totalExchangerTx:
        description: Total number of exchanger  transactions
        type: integer
      totalNFT:
        description: Total number of NFTs
        type: integer
      totalNFTAmount:
        description: The total amount of coins in the chain
        type: string
      totalNFTCollection:
        description: Total number of NFT collections
        type: integer
      totalNFTCreator:
        description: Total creator of NFTs
        type: integer
      totalNFTTx:
        description: Total number of  NFT transactions
        type: integer
      totalRecycle:
        description: Total number of recycle SNFT
        type: integer
      totalSNFT:
        description: Total number of SNFTs
        type: integer
      totalSNFTAmount:
        description: The total amount of coins in the chain
        type: string
      totalSNFTCollection:
        description: Total number of SNFT collections
        type: integer
      totalSNFTCreator:
        description: Total creator of SNFTs
        type: integer
      totalSNFTTx:
        description: Total number of  SNFT transactions
        type: integer
      totalTransaction:
        description: Total number of transactions
        type: integer
      totalUncle:
        description: Number of total uncle blocks
        type: integer
    type: object
  service.CollectionsRes:
    properties:
      collections:
        description: NFT collection list
        items:
          $ref: '#/definitions/model.Collection'
        type: array
      total:
        description: The total number of NFT collections
        type: integer
    type: object
...
  service.ErrRes:
    properties:
      err_str:
        description: Error message
        type: string
    type: object
...
  service.LineChartRes:
    properties:
      blocks:
        items:
          properties:
            number:
              type: integer
            txCount:
              type: integer
          type: object
        type: array
      exchangers:
        items:
          properties:
            day:
              type: string
            index:
              type: integer
            num:
              type: integer
          type: object
        type: array
      txs:
        items:
          properties:
            gasPrice:
              type: integer
            hash:
              type: string
          type: object
        type: array
    type: object
...
info:
  contact: {}
  description: Block browser back-end interface, parses data from the blockchain,
    provides information retrieval services for blocks, transactions, NFT, SNFT, NFT
    collections, and exchanges
  title: block explorer API
  version: "1.0"
paths:
  /account/{addr}:
    get:
      consumes:
      - application/json
      description: Query the account information of the specified address
      parameters:
      - description: account address
        in: path
        name: addr
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AccountRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query one account
      tags:
      - account
  /account/page:
    get:
      consumes:
      - application/json
      description: set the account ranking according to the amount of coins held
      parameters:
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AccountsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query top accounts
      tags:
      - account
  /block/{number}:
    get:
      consumes:
      - application/json
      description: specifies the number query block
      parameters:
      - description: Block number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Block'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query block
      tags:
      - block
...
  /block/page:
    get:
      consumes:
      - application/json
      description: Query the block list in reverse order of height
      parameters:
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.BlocksRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query block list
      tags:
      - block
...
  /chart/line:
    get:
      consumes:
      - application/json
      description: query charts
      parameters:
      - description: Limit, default 10
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LineChartRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query charts
      tags:
      - chart
  /collection/{id}:
    get:
      consumes:
      - application/json
      description: specifies the ID to query the NFT collection information
      parameters:
      - description: Collection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query NFT collection
      tags:
      - NFT Collection
  /collection/page:
    get:
      consumes:
      - application/json
      description: Query NFT collection list in reverse order of created block height
      parameters:
      - description: Exchange, if empty, query all exchanges
        in: query
        name: exchanger
        type: string
      - description: Creator, if empty, query all
        in: query
        name: creator
        type: string
      - description: 'Type: all, nft, snft, default all'
        in: query
        name: type
        type: string
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.CollectionsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: Query the list of NFT collections
      tags:
      - NFT Collection
...
  /totals:
    get:
      consumes:
      - application/json
      description: Query the total number of blocks, transactions, accounts, etc.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Cache'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query some total data
      tags:
      - block
...
swagger: "2.0"